version2/requirements.txt
sentence-transformers
faiss-cpu
PyPDF2
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
llama-cpp-python
pyyaml
requests
beautifulsoup4
pytesseract
pillow
pdf2image
